fixed HTML tags
